Refactor: move updateCache() for re-use
