+++ src/main/java/hudson/plugins/tfs/commands/IsWorkspaceMappedCommand.java
@@ -1,11 +1,7 @@
 package hudson.plugins.tfs.commands;
 
 import com.microsoft.tfs.core.TFSTeamProjectCollection;
-import com.microsoft.tfs.core.clients.versioncontrol.VersionControlClient;
-import com.microsoft.tfs.core.clients.versioncontrol.VersionControlConstants;
-import com.microsoft.tfs.core.clients.versioncontrol.Workstation;
 import com.microsoft.tfs.core.clients.versioncontrol.soapextensions.Workspace;
-import com.microsoft.tfs.core.config.persistence.PersistenceStoreProvider;
 import hudson.model.TaskListener;
 import hudson.plugins.tfs.model.MockableVersionControlClient;
 import hudson.plugins.tfs.model.Server;
@@ -48,10 +44,4 @@ public class IsWorkspaceMappedCommand extends AbstractCallableCommand implements
         return existsMapping;
     }
 
-    static void updateCache(final TFSTeamProjectCollection connection) {
-        final PersistenceStoreProvider persistenceStoreProvider = connection.getPersistenceStoreProvider();
-        final Workstation workstation = Workstation.getCurrent(persistenceStoreProvider);
-        final VersionControlClient vcc = connection.getVersionControlClient();
-        workstation.updateWorkspaceInfoCache(vcc, VersionControlConstants.AUTHENTICATED_USER);
-    }
 }
